+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Also: [Deployment history](https://github.com/hippware/tr-wiki/wiki/Client-deplo
 
 Ticket numbers refer to the ticket tracker for this project if not specified.
 
+* Expressing user activity via color/faded avatar (Rework #3658)
+  * (PR #3788, PR #3791)
+* Better fade header nav config (PR #3787)
+  * Back caret missing on friends search (#3738)
+  * Missing Header on Location Profile (#3748)
+* Split out bicycle and running for user activity (#3756)
+* Traffic dumps for client side tests (PR #3790)
+
 # 4.13.0 - 2019 May 30
 
 * Red boxes in top left corner of map (#3757)